Updated v1.0 patch format
Updated v1.0 patches from git 2.34.1 format to 2.45.2 format. The
update is mostly superficial, but it includes a relevant change to the
hash length that ensures future collision avoidance.

This also fixes a very minor inconsistency in the cover letter git
version delimteter. There is supposed to be a space " " after the
delimiter, as in "-- ", but that was missing on the cover letter.